+++ lib/command_line.rb
@@ -1,0 +1,11 @@
+class CommandLine
+  attr_accessor :command
+
+  def initialize(command)
+    @command = command
+  end
+
+  def result
+    IO.popen("#{@command}") { |answer| answer.read.chomp }
+  end
+end
