+++ lib/data/ls_commands.txt
@@ -1,3 +1,4 @@
 ls - список файлов и каталогов в текущем каталоге
 pwd - путь к текущей дериктории или файлу
 cat your_file - просмотр содержимого файла
+exit - выход
